Moved and Deleted Redundant Files
