+++ vcell-cli/src/main/java/org/vcell/cli/commands/conversion/ExportOmexBatchCommand.java
@@ -1,9 +1,10 @@
-package org.vcell.cli.vcml;
+package org.vcell.cli.commands.conversion;
 
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.core.LoggerContext;
+import org.vcell.cli.vcml.VcmlOmexConverter;
 import org.vcell.trace.Tracer;
 import org.vcell.sedml.ModelFormat;
 import picocli.CommandLine.Command;
