+++ vcell-cli/src/main/java/org/vcell/cli/messaging/CLIRecordable.java
@@ -1,4 +1,4 @@
-package org.vcell.cli;
+package org.vcell.cli.messaging;
 
 import java.io.IOException;
