+++ vcell-cli/src/main/java/org/vcell/cli/run/ExecutionJob.java
@@ -3,7 +3,7 @@ package org.vcell.cli.run;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.vcell.cli.CLIRecordable;
+import org.vcell.cli.messaging.CLIRecordable;
 import org.vcell.cli.exceptions.ExecutionException;
 import org.vcell.cli.exceptions.PreProcessingException;
 import org.vcell.cli.run.hdf5.BiosimulationsHdf5Writer;
