+++ vcell-cli/src/main/java/org/vcell/cli/run/SedmlJob.java
@@ -7,7 +7,7 @@ import org.apache.commons.lang.SystemUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jlibsedml.*;
-import org.vcell.cli.CLIRecordable;
+import org.vcell.cli.messaging.CLIRecordable;
 import org.vcell.cli.exceptions.ExecutionException;
 import org.vcell.cli.exceptions.PreProcessingException;
 import org.vcell.cli.run.hdf5.HDF5ExecutionResults;
